+++ src/main.cc
@@ -5,6 +5,7 @@
 
 #include "reduct/api/api_server.h"
 #include "reduct/async/loop.h"
+#include "reduct/auth/token_auth.h"
 #include "reduct/config.h"
 #include "reduct/core/env_variable.h"
 #include "reduct/core/logger.h"
@@ -12,6 +13,7 @@
 
 using reduct::api::IApiServer;
 using reduct::async::ILoop;
+using reduct::auth::ITokenAuthentication;
 using reduct::core::EnvVariable;
 using reduct::core::Error;
 using reduct::core::Logger;
@@ -37,22 +39,21 @@ int main() {
   auto port = env.Get<int>("RS_PORT", 8383);
   auto api_base_path = env.Get<std::string>("RS_API_BASE_PATH", "/");
   auto data_path = env.Get<std::string>("RS_DATA_PATH", "/data");
+  auto api_token = env.Get<std::string>("RS_API_TOKEN", "", true);
 
   Logger::set_level(log_level);
 
   LOG_INFO("Configuration: \n {}", env.Print());
 
-  auto storage = IStorage::Build({
-      .data_path = data_path,
-  });
-
   Loop loop;
   ILoop::set_loop(&loop);
-  auto server = IApiServer::Build(std::move(storage), {
-                                                          .host = host,
-                                                          .port = port,
-                                                          .base_path = api_base_path,
-                                                      });
+
+  IApiServer::Components components{
+      .handler = IStorage::Build({.data_path = data_path}),
+      .auth = ITokenAuthentication::Build(api_token),
+  };
+
+  auto server = IApiServer::Build(std::move(components), {.host = host, .port = port, .base_path = api_base_path});
   server->Run(running);
 
   return 0;