+++ src/reduct/api/callbacks.h
@@ -7,18 +7,12 @@
 
 #include "reduct/async/run.h"
 #include "reduct/core/error.h"
-#include "reduct/proto/api/bucket_info.pb.h"
-#include "reduct/proto/api/bucket_settings.pb.h"
-#include "reduct/proto/api/server_info.pb.h"
+#include "reduct/core/result.h"
+#include "reduct/proto/api/auth.pb.h"
+#include "reduct/proto/api/bucket.pb.h"
+#include "reduct/proto/api/server.pb.h"
+
 namespace reduct::api {
-
-template <typename Response>
-struct CallbackResult {
-  Response response;
-  core::Error error;
-
-  operator const core::Error&() const { return error; }
-};
 
 //---------------------
 // Server API
@@ -33,7 +27,7 @@ class IInfoCallback {
     proto::api::ServerInfo info;
   };
   struct Request {};
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnInfo(const Request& req) const = 0;
 };
 
@@ -46,8 +40,20 @@ class IListStorageCallback {
     proto::api::BucketInfoList buckets;
   };
   struct Request {};
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnStorageList(const Request& req) const = 0;
+};
+
+//---------------------
+// Auth API
+//---------------------
+class IRefreshToken {
+ public:
+  using Request = std::string;
+  using Response = proto::api::RefreshTokenResponse;
+
+  using Result = core::Result<Response>;
+  virtual async::Run<Result> OnRefreshToken(const Request& req) const = 0;
 };
 
 //---------------------
@@ -64,7 +70,7 @@ class ICreateBucketCallback {
     std::string_view bucket_name;
     proto::api::BucketSettings bucket_settings;
   };
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnCreateBucket(const Request& req) = 0;
 };
 
@@ -80,7 +86,7 @@ class IGetBucketCallback {
   struct Request {
     std::string_view bucket_name;
   };
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnGetBucket(const Request& req) const = 0;
 };
 
@@ -93,7 +99,7 @@ class IRemoveBucketCallback {
   struct Request {
     std::string_view bucket_name;
   };
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnRemoveBucket(const Request& req) = 0;
 };
 
@@ -108,7 +114,7 @@ class IUpdateBucketCallback {
     proto::api::BucketSettings new_settings;
   };
 
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnUpdateCallback(const Request& req) = 0;
 };
 //---------------------
@@ -127,7 +133,7 @@ class IWriteEntryCallback {
     std::string_view blob;
   };
 
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnWriteEntry(const Request& req) = 0;
 };
 
@@ -146,7 +152,7 @@ class IReadEntryCallback {
     std::string_view timestamp;
   };
 
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   virtual async::Run<Result> OnReadEntry(const Request& req) = 0;
 };
 
@@ -168,7 +174,7 @@ class IListEntryCallback {
     std::string_view stop_timestamp;
   };
 
-  using Result = CallbackResult<Response>;
+  using Result = core::Result<Response>;
   [[nodiscard]] virtual async::Run<Result> OnListEntry(const Request& req) const = 0;
 };