+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * PUT method of Bucket API has optional parameters, [PR-32](https://github.com/reduct-storage/reduct-storage/pull/32)
 * Extends information from GET method of Server API, [PR-33](https://github.com/reduct-storage/reduct-storage/pull/33)
 * Add GET /list end point to browse buckets, [PR-34](https://github.com/reduct-storage/reduct-storage/pull/34)
+* Add bearer token authentication, [PR-36](https://github.com/reduct-storage/reduct-storage/pull/36)
 
 **Bugs**: